+++ resources/lang/en/validation.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * ----------------------------------------------------------
+ * Create all of the validation error messages
+ * ----------------------------------------------------------
+ */
+return [
+    "email" => "The :field must be a valid email address",
+    "equals" => "The :field field did not match the expected value",
+    "required" => "The :field field is required"
+];